+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- Export content and write to file using a generator/yield. This avoids memory ballooning to the size of the exported file. [fredvd]
+- Export portlets.
+  [pbauer]
+
+- Export content and write to file using a generator/yield. This avoids memory ballooning to the size of the exported file. 
+  [fredvd]
 
 
 1.1 (2021-08-02)